Hotel room manager web tool
Tool that would work for a small hotel to manage its rooms.
The user can Inser,Search one or get the entire list of the rooms written on the database.
The db is stanced on firebase,and the backend can be found in the hrm-functions repo,consisted in cloud functions to execute requests.